MEDWRITES
Healthcare is one from the metaphorical Stone Age. The workflow, with doctors, pharmacies, health insurance companies, patients and, of course, the medical prescription itself playing the main roles, comes to a standstill. If another factor is added at the same time, for example in the form of a pandemic, shipwreck is inevitable. In order to successfully integrate digitization into the healthcare system, medical prescriptions in digital form are an innovative necessity.
                       Physicians use EMR software to enter all details about the patient. They also enter the medications prescribed, i.e., enter the prescription. Entry of a prescription in a software takes much more time than hand written prescription, but data entry in software enables prescription data to be available to pharmacy and insurance companies, and lends itself to research & analytics. The data entry burden of prescriptions should be reduced, and be comparable to that of hand written prescription. There is a requirement of a system to be able to enter a prescription with structured data in a software system, within a comparable time to hand written prescription